Move register eval hooks from deployer to mastra init
